Split typechecker restored signature tests
