+++ src/typechecker/tests/resolver_collection/behavior_methods/restored_signatures.rs
@@ -1,287 +1,5 @@
 use super::*;
 
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_method_metadata() {
-    let mut program = parse_program(
-        r#"
-Mapper: behavior {
-    map: (Self, (i32) i32) (i32) i32
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::Behavior { methods, .. } = &mut program.declarations[0] {
-        methods[0].params[1].ty = AstType::I32;
-        methods[0].return_type = Some(AstType::I32);
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let info = tc.behaviors.get("Mapper").expect("behavior info");
-    assert_eq!(
-        info.methods[0].params[1].ty,
-        AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::I32),
-        }
-    );
-    assert_eq!(
-        info.methods[0].return_type,
-        Some(AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::I32),
-        })
-    );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_method_name_metadata() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) StaticString
-}
-
-Point.implements(Json) {
-    encode = (value: Point) StaticString { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::Behavior { methods, .. } = &mut program.declarations[1] {
-        methods[0].name = "missing".to_string();
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let info = tc.behaviors.get("Json").expect("behavior info");
-    assert_eq!(info.methods[0].name, "encode");
-    assert!(
-            tc.diagnostics.is_empty(),
-            "resolver-restored behavior method name metadata should avoid stale AST impl diagnostics: {:?}",
-            tc.diagnostics
-        );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_method_return_presence_metadata() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) StaticString
-}
-
-Point.implements(Json) {
-    encode = (value: Point) StaticString { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::Behavior { methods, .. } = &mut program.declarations[1] {
-        methods[0].return_type = None;
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let info = tc.behaviors.get("Json").expect("behavior info");
-    assert_eq!(info.methods[0].return_type, Some(AstType::Str));
-    assert!(
-            tc.diagnostics.is_empty(),
-            "resolver-restored behavior method return metadata should avoid stale AST impl diagnostics: {:?}",
-            tc.diagnostics
-        );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_method_parameter_count() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) StaticString
-}
-
-Point.implements(Json) {
-    encode = (value: Point) StaticString { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::Behavior { methods, .. } = &mut program.declarations[1] {
-        methods[0].params.push(Param {
-            name: "stale".to_string(),
-            ty: AstType::I32,
-            mutable: false,
-            span: Span::dummy(),
-        });
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let info = tc.behaviors.get("Json").expect("behavior info");
-    assert_eq!(info.methods[0].params.len(), 1);
-    assert_eq!(info.methods[0].params[0].ty, AstType::SelfType);
-    assert!(
-        tc.diagnostics.is_empty(),
-        "resolver-restored behavior method params should avoid stale AST impl diagnostics: {:?}",
-        tc.diagnostics
-    );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_method_missing_parameter_count() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Mapper: behavior {
-    map: (Self, i32) StaticString
-}
-
-Point.implements(Mapper) {
-    map = (value: Point, input: i32) StaticString { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::Behavior { methods, .. } = &mut program.declarations[1] {
-        methods[0].params.pop();
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let info = tc.behaviors.get("Mapper").expect("behavior info");
-    assert_eq!(info.methods[0].params.len(), 2);
-    assert_eq!(info.methods[0].params[0].name, "__arg0");
-    assert_eq!(info.methods[0].params[1].name, "__arg1");
-    assert_eq!(info.methods[0].params[0].ty, AstType::SelfType);
-    assert_eq!(info.methods[0].params[1].ty, AstType::I32);
-    assert!(
-        tc.diagnostics.is_empty(),
-        "resolver-restored behavior method params should avoid stale AST impl diagnostics: {:?}",
-        tc.diagnostics
-    );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_method_parameter_names() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (value: Self) StaticString
-}
-
-Point.implements(Json) {
-    encode = (value: Point) StaticString { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::Behavior { methods, .. } = &mut program.declarations[1] {
-        methods[0].params[0].name = "stale".to_string();
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let info = tc.behaviors.get("Json").expect("behavior info");
-    assert_eq!(info.methods[0].params[0].name, "value");
-    assert_eq!(info.methods[0].params[0].ty, AstType::SelfType);
-    assert!(
-            tc.diagnostics.is_empty(),
-            "resolver-restored behavior method parameter names should avoid stale AST impl diagnostics: {:?}",
-            tc.diagnostics
-        );
-}
-
-#[test]
-fn collect_declarations_with_symbols_ignores_stale_behavior_method_parameter_order() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Mapper: behavior {
-    map: (value: Self, input: i32) StaticString
-}
-
-Point.implements(Mapper) {
-    map = (value: Point, input: i32) StaticString { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::Behavior { methods, .. } = &mut program.declarations[1] {
-        methods[0].params.swap(0, 1);
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let info = tc.behaviors.get("Mapper").expect("behavior info");
-    assert_eq!(info.methods[0].params[0].name, "value");
-    assert_eq!(info.methods[0].params[1].name, "input");
-    assert_eq!(info.methods[0].params[0].ty, AstType::SelfType);
-    assert_eq!(info.methods[0].params[1].ty, AstType::I32);
-    assert!(
-            tc.diagnostics.is_empty(),
-            "resolver-restored behavior method parameter order should avoid stale AST impl diagnostics: {:?}",
-            tc.diagnostics
-        );
-}
-
-#[test]
-fn collect_declarations_with_symbols_uses_resolver_behavior_method_count() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) StaticString
-    describe: (Self) StaticString
-}
-
-Point.implements(Json) {
-    encode = (value: Point) StaticString { "point" }
-    describe = (value: Point) StaticString { "desc" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    if let Declaration::Behavior { methods, .. } = &mut program.declarations[1] {
-        methods.pop();
-    }
-    let mut tc = TypeChecker::new();
-
-    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
-
-    let info = tc.behaviors.get("Json").expect("behavior info");
-    assert_eq!(info.methods.len(), 2);
-    assert_eq!(info.methods[0].name, "encode");
-    assert_eq!(info.methods[1].name, "describe");
-    assert!(
-        tc.diagnostics.is_empty(),
-        "resolver-restored behavior methods should avoid stale AST impl diagnostics: {:?}",
-        tc.diagnostics
-    );
-}
+mod method_shapes;
+mod names_and_returns;
+mod parameter_metadata;
